serde_flexitos
Module de

DeserializeSeed and Visitor impls for deserializing trait objects and collections of trait objects.

Structs

DeserializeMapWith

Deserialize HashMap<K, V>, using key_deserialize_seed to deserialize K, and value_deserialize_seed to deserialize V. Implements DeserializeSeed. Use the following functions to create instances of this struct:

DeserializeTraitObject

Deserialize Box<<R as Registry>::TraitObject> from a single id-value pair, using the registry to get deserialize functions for concrete types of the trait object. Implements DeserializeSeed.

DeserializeVecWithTraitObject

Deserialize Vec<Box<<R as Registry>::TraitObject>>, using the registry to get deserialize functions for concrete types of the trait object. Implements DeserializeSeed.