Struct serde_flexitos::de::DeserializeMapWith

pub struct DeserializeMapWith<K, V> { /* private fields */ }

Deserialize HashMap<K, V>, using key_deserialize_seed to deserialize K, and value_deserialize_seed to deserialize V. Implements DeserializeSeed. Use the following functions to create instances of this struct:

trait_object_key: deserialize map keys as trait objects,

trait_object_value: deserialize map values as trait objects,

trait_object_key_and_value: deserialize map keys and values as trait objects.

Implementations

impl<'k, K, V, R> DeserializeMapWith<DeserializeTraitObject<'k, R>, PhantomData<V>>
where K: Eq + Hash + ?Sized, R: Registry<TraitObject = K>,

pub fn trait_object_key(registry: &'k R) -> Self

Deserialize HashMap<Box<K>, V>, deserializing Box<K> as a trait object where K is the trait object type, using registry to get deserialize functions for concrete types of trait object K.

impl<'v, K, V, R> DeserializeMapWith<PhantomData<K>, DeserializeTraitObject<'v, R>>
where K: Eq + Hash, V: ?Sized, R: Registry<TraitObject = V>,

pub fn trait_object_value(registry: &'v R) -> Self

Deserialize HashMap<K, Box<V>>, deserializing Box<V> as a trait object where V is the trait object type, using registry to get deserialize functions for concrete types of trait object V.

impl<'k, 'v, K, RK, V, RV> DeserializeMapWith<DeserializeTraitObject<'k, RK>, DeserializeTraitObject<'v, RV>>
where K: Eq + Hash + ?Sized, V: ?Sized, RK: Registry<TraitObject = K>, RV: Registry<TraitObject = V>,

pub fn trait_object_key_and_value( key_registry: &'k RK, value_registry: &'v RV ) -> Self

Deserialize HashMap<Box<K>, Box<V>>:

deserialize Box<K> as a trait object where K is the trait object type, using key_registry to get deserialize functions for concrete types of trait object K.

deserialize Box<V> as a trait object where V is the trait object type, using value_registry to get deserialize functions for concrete types of trait object V.