Trait serde_flexitos::Registry

pub trait Registry {
    type Identifier;
    type TraitObject: ?Sized;

    // Required methods
    fn register(
        &mut self,
        id: Self::Identifier,
        deserialize_fn: DeserializeFn<Self::TraitObject>,
    );
    fn get_deserialize_fn(
        &self,
        id: Self::Identifier,
    ) -> Result<&DeserializeFn<Self::TraitObject>, GetError<Self::Identifier>>;
    fn get_trait_object_name(&self) -> &'static str;

    // Provided methods
    fn register_type<T>(&mut self, id: Self::Identifier)
       where T: DeserializeOwned + Into<Box<Self::TraitObject>> { ... }
    fn deserialize_trait_object<'de, D>(
        &self,
        deserializer: D,
    ) -> Result<Box<Self::TraitObject>, D::Error>
       where D: Deserializer<'de>,
             Self: Sized,
             Self::Identifier: Deserialize<'de> + Debug { ... }
}

Registry mapping unique identifiers of types to their deserialize implementations, enabling deserialization of a specific trait object type.

Required Associated Types

type Identifier

The type of unique identifiers this registry uses. &'static str is used as a default.

type TraitObject: ?Sized

The trait object type this registry maps deserialize functions for.

Required Methods

fn register( &mut self, id: Self::Identifier, deserialize_fn: DeserializeFn<Self::TraitObject>, )

Register deserialize_fn as the deserialize function for id.

If id was already registered before, get_deserialize_fn will forever return Err(MultipleRegistrations) for that id.

fn get_deserialize_fn( &self, id: Self::Identifier, ) -> Result<&DeserializeFn<Self::TraitObject>, GetError<Self::Identifier>>

Gets the deserialize function for id.

Errors

GetError::NotRegistered { id } if no deserialize function was registered for id.

GetError::MultipleRegistrations { id } if multiple deserialize functions were registered for id.

fn get_trait_object_name(&self) -> &'static str

Gets the trait object name, for diagnostic purposes.

Provided Methods

fn register_type<T>(&mut self, id: Self::Identifier)
where T: DeserializeOwned + Into<Box<Self::TraitObject>>,

Register a default deserialize function for type T as the deserialize function for id. T must implement DeserializeOwned and must be convertable into Box<Self::TraitObject> with Into<Box<Self::TraitObject>>.

If id was already registered before, get_deserialize_fn will forever return Err(MultipleRegistrations) for that id.

fn deserialize_trait_object<'de, D>( &self, deserializer: D, ) -> Result<Box<Self::TraitObject>, D::Error>
where D: Deserializer<'de>, Self: Sized, Self::Identifier: Deserialize<'de> + Debug,

Deserialize a trait object with deserializer, using this registry to get the deserialize function for the concrete type, based on the deserialized ID.

Errors

Returns an error if 0 or more than 1 deserialize functions were registered for the deserialized ID.

Object Safety

This trait is not object safe.